% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{blankErr}
\alias{blankErr}
\title{Determine error in blankfm}
\usage{
blankErr(blankfm)
}
\arguments{
\item{blankfm}{A vector of normalized Fm's of blanks.}
}
\value{
An error for the given blanks.
}
\description{
Uses the "error floor" method from SNICSer
}
