% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{calcRaw}
\alias{calcRaw}
\title{Calculate per-block fields}
\usage{
calcRaw(data)
}
\arguments{
\item{data}{A data table of raw AMS data}
}
\value{
The data table with fields for corrected 14/12,
internal error, and d13C added
}
\description{
Calculate per-block fields
}
