% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{checkDB}
\alias{checkDB}
\title{Check odbc call returns for errors}
\usage{
checkDB(data)
}
\arguments{
\item{data}{The return from a call to sqlQuery}
}
\description{
Check odbc call returns for errors
}
