% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{conNOSAMS}
\alias{conNOSAMS}
\title{Open NOSAMS DB connection}
\usage{
conNOSAMS()
}
\value{
A RODBC db connection object
}
\description{
Takes a connection string from the CONSTRING environment variable.
}

