% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micadas_db.R
\name{convert_micadas_snics}
\alias{convert_micadas_snics}
\title{Convert MICADAS results to SNICS results format}
\usage{
convert_micadas_snics(micadas_df)
}
\arguments{
\item{micadas_df}{MICADAS data in `get_magazine()` format.}
}
\value{
A dataframe in SNICS Results format
}
\description{
Convert MICADAS results to SNICS results format
}
\seealso{
\code{\link{get_magazine}} to store user credentials
in the system key store, and \code{\link{write_snics_results}} for
writing data to a SNICS results file.
}
