% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{d14c}
\alias{d14c}
\title{Calculate D14C.}
\usage{
d14c(fm, yc)
}
\arguments{
\item{fm}{Fraction modern.}

\item{yc}{Year of collection.}
}
\value{
A vector of D14C values
}
\description{
Calculate D14C given fraction modern and year of collection
}
