% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{dctofm}
\alias{dctofm}
\title{Calculate Fm}
\usage{
dctofm(dc, yc)
}
\arguments{
\item{dc}{D14C of sample.}

\item{yc}{Year of collection.}
}
\value{
Fraction modern of sample.
}
\description{
Calculate Fm from D14C
}
