% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{doCor1412}
\alias{doCor1412}
\title{Correct raw 14/12 for fractionation using 13/12 ratio}
\usage{
doCor1412(he1412, he1312)
}
\arguments{
\item{he1412}{Raw measured 14/12}

\item{he1312}{Raw measured 13/12}
}
\value{

}
\description{
Correct raw 14/12 for fractionation using 13/12 ratio
}
