% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{doMBCerr}
\alias{doMBCerr}
\title{Propagate mass balance blank correction error}
\usage{
doMBCerr(
  fmmeas,
  fmblank,
  massmeas,
  massblank,
  fmmeaserr,
  fmblankerr,
  massmeaserr,
  massblankerr
)
}
\arguments{
\item{fmmeas}{Normalized Fm of the sample}

\item{fmblank}{Normalized Fm of the blank}

\item{massmeas}{Mass of the sample}

\item{massblank}{Mass of the blank}

\item{fmmeaserr}{Measurement error of the sample}

\item{fmblankerr}{Measurement error of the blank}

\item{massmeaserr}{Mass error of the sample}

\item{massblankerr}{Mass error of the blank}
}
\value{

}
\description{
Note, use total mass - blank mass for mass of the sample
if appropriate
}
