% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.R
\name{formatQC}
\alias{formatQC}
\title{Format QC Table to match standards table format}
\usage{
formatQC(data)
}
\arguments{
\item{data}{A data frame as returned by getQCStandards.}
}
\description{
Format QC Table to match standards table format
}

