% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micadas_db.R
\name{get_bats_results}
\alias{get_bats_results}
\title{Get BATS results from MICADAS DB}
\usage{
get_bats_results(magazine)
}
\arguments{
\item{magazine}{Name of MICADAS magazine}
}
\value{
A dataframe of BATS results
}
\description{
Get BATS results from MICADAS DB
}
