% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micadas_db.R
\name{get_magazine}
\alias{get_magazine}
\title{Get run data for a list of magazine names.}
\usage{
get_magazine(magazine)
}
\arguments{
\item{magazine}{A vector of magazine names in character format}
}
\value{
A data frame of raw data
}
\description{
Get run data for a list of magazine names.
}
