% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getdc13}
\alias{getdc13}
\title{get info from dc13 table}
\usage{
getdc13(tps)
}
\arguments{
\item{tps}{A vector of tp_nums}
}
\value{
A data frame of matching rows from dc13
}
\description{
get info from dc13 table
}
