% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getOStandards}
\alias{getOStandards}
\title{Get Old Standards}
\usage{
getOStandards(from, to = "present", sys = "both", rec = NULL,
  osg = NULL)
}
\arguments{
\item{from}{Character vector of date in form 'YYYY-MM-DD'.}

\item{to}{Character vector of date in form 'YYYY-MM-DD'.}

\item{sys}{Character vector of system name: 'cfams', 'usams', or 'both'.}

\item{rec}{Numeric vector of NOSAMS reciept numbers. If supplied, get
all records matching these receipt numbers.}

\item{osg}{Numeric vector of NOSAMS osg numbers. Supersedes 'rec' parameter.}
}
\value{
A data frame of standards data
}
\description{
Get standards from database using standards table.
}
