% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getProcess}
\alias{getProcess}
\title{Get the process id for a target}
\usage{
getProcess(tp_num, .con = con)
}
\arguments{
\item{tp_num}{The numeric target identifier}

\item{.con}{A odbc database connection The numeric target identifier}
}
\value{
A short character code for the process
}
\description{
Get the process id for a target
}
