% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getProcessDesc}
\alias{getProcessDesc}
\title{Get the process description for a target}
\usage{
getProcessDesc(tp_num, .con = con)
}
\arguments{
\item{tp_num}{The numeric target identifier}

\item{.con}{A odbc database connection The numeric target identifier}
}
\value{
A short character description for the process
}
\description{
Get the process description for a target
}
