% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.R
\name{getQCData}
\alias{getQCData}
\title{Get and format standards data}
\usage{
getQCData(from, to = "present", sys = "both", useQC = FALSE,
  intcal = TRUE)
}
\arguments{
\item{from}{Character vector of date in form 'YYYY-MM-DD'.}

\item{to}{Character vector of date in form 'YYYY-MM-DD'.}

\item{sys}{Character vector of system name: 'cfams', 'usams', or 'both'.}

\item{useQC}{Use data from QC table if true.}
}
\description{
Get and format standards data
}

