% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getQCTable}
\alias{getQCTable}
\title{Get secondary data from qc table}
\usage{
getQCTable(from, to, sys = "both")
}
\arguments{
\item{from}{Character vector of date in form 'YYYY-MM-DD'.}

\item{to}{Character vector of date in form 'YYYY-MM-DD'.}

\item{sys}{Character vector of system name: 'cfams', 'usams', or 'both'.}
}
\value{
A data frame of qc table data
}
\description{
Get secondary data from qc table
}
