% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getRawData}
\alias{getRawData}
\title{Get raw ams data from database}
\usage{
getRawData(from, to)
}
\arguments{
\item{from}{A date in character form (m-d-Y)}

\item{to}{A date in character. Defaults to present.}
}
\value{
A data frame of raw data
}
\description{
Get raw ams data from database
}
