% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getRawWheel}
\alias{getRawWheel}
\title{Get Raw Wheel}
\usage{
getRawWheel(wheel)
}
\arguments{
\item{wheel}{A vector of wheel names in character format}
}
\value{
A data frame of raw data
}
\description{
Get Raw Wheel
}
