% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getRecOS}
\alias{getRecOS}
\title{Get results by recnum from os}
\usage{
getRecOS(recnum)
}
\arguments{
\item{recnum}{A vector of recnums}
}
\value{
A dataframe of summary statistics
}
\description{
Get results by recnum from os
}
