% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getRecSR}
\alias{getRecSR}
\title{Get results by recnum from snics_results}
\usage{
getRecSR(recnum)
}
\arguments{
\item{recnum}{A vector of recnums}
}
\value{
A dataframe of summary statistics
}
\description{
Get results by recnum from snics_results
}
