% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getTPnum}
\alias{getTPnum}
\title{Get results by tpnum from snics_results}
\usage{
getTPnum(tpnum)
}
\arguments{
\item{tpnum}{A vector of tpnums}
}
\value{
A dataframe of summary statistics
}
\description{
Get results by tpnum from snics_results
}
