% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getWheel}
\alias{getWheel}
\title{Get results for a list of wheel names.}
\usage{
getWheel(wheel)
}
\arguments{
\item{wheel}{A vector of wheel names in character format}
}
\value{
A data frame of analysed data
}
\description{
Get results for a list of wheel names.
}
