% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getWheelInfo}
\alias{getWheelInfo}
\title{Get info for a wheel}
\usage{
getWheelInfo(wheel)
}
\arguments{
\item{wheel}{Character vector of wheel name in form '[CF|US]AMSMMDDYY'.}
}
\value{
A data frame of wheel information
}
\description{
Get info for a wheel
}
