% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getWheelStandards}
\alias{getWheelStandards}
\title{Get standards from a wheel}
\usage{
getWheelStandards(wheel)
}
\arguments{
\item{wheel}{}
}
\value{
A dataframe of summary statistics
}
\description{
Get standards from a wheel
}
