% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getWheelStdsSR}
\alias{getWheelStdsSR}
\title{Get standards from a wheel from snics_results}
\usage{
getWheelStdsSR(wheel)
}
\arguments{
\item{wheel}{A wheelname as a character vector}
}
\value{
A dataframe of summary statistics
}
\description{
Get standards from a wheel from snics_results
}
