% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{intrErr}
\alias{intrErr}
\title{Intrinsic Error}
\usage{
intrErr(popErr, targErr)
}
\arguments{
\item{popErr}{Standard deviation of the population.}

\item{targErr}{Measurement error of a measurement.}
}
\description{
Calculate the "extra" (intrinsic, residual) error.
}

