% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micadas_db.R
\name{make_micadas_resultsfile}
\alias{make_micadas_resultsfile}
\title{Get MICADAS results and write SNICS resultsfile}
\usage{
make_micadas_resultsfile(magazine, filename)
}
\arguments{
\item{filename}{}
}
\value{
Silently returns a dataframe of MICADAS data
}
\description{
Get MICADAS results and write SNICS resultsfile
}
