% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{meanRat}
\alias{meanRat}
\title{Mean ratio for a target, weighted by error}
\usage{
meanRat(normRat, normRatErr)
}
\arguments{
\item{normRat}{A vector of normalized ratios.}

\item{normRatErr}{A vector of errors in measurements of the ratio.}
}
\value{
The weighted mean ratio for the target.
}
\description{
Mean ratio for a target, weighted by error
}
