% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.R
\name{mungeQCTable}
\alias{mungeQCTable}
\title{Modify and calculate fields for raw qc data.}
\usage{
mungeQCTable(data)
}
\arguments{
\item{data}{A data frame as returned by getQCTable}
}
\value{
A data frame with calculated fields.
}
\description{
Modify and calculate fields for raw qc data.
}
