% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{mungeResfile}
\alias{mungeResfile}
\title{Process wheel results data frame.}
\usage{
mungeResfile(data)
}
\arguments{
\item{data}{A data table as returned by readResfile.}
}
\value{
A results data table with fixed and calculated fields.
}
\description{
Process wheel results data frame.
}
