% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.R
\name{mungeStandards}
\alias{mungeStandards}
\title{Modify and calculate fields for raw standards data.}
\usage{
mungeStandards(data, std)
}
\arguments{
\item{data}{A data frame as returned by getStandards}

\item{std}{The standards table as returned by getStdTable}
}
\value{
A data frame of standards with calculated fields.
}
\description{
Modify and calculate fields for raw standards data.
}
