% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{normFm}
\alias{normFm}
\title{Normalized Value}
\usage{
normFm(fmm, fmc)
}
\arguments{
\item{fmm}{Measured value.}

\item{fmc}{Expected value.}
}
\value{
A vector of normalized fm
}
\description{
Normalize a value using the expected value or population mean.
}
