% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{normRun}
\alias{normRun}
\title{Normalize a run using the measured and consensus value of a standard}
\usage{
normRun(sample, standard, stdrat = 1.0398)
}
\arguments{
\item{sample}{Ratio to be normalized.}

\item{standard}{Measured ratio of the standard.}

\item{stdrat}{Consensus value of the standard.}
}
\value{
The normalized ratio of the sample.
}
\description{
Normalize a run using the measured and consensus value of a standard
}
