% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{normRunExtErr}
\alias{normRunExtErr}
\title{Calculate external (deviation of runs) error for a target}
\usage{
normRunExtErr(normRat)
}
\arguments{
\item{normRat}{A vector of normalized ratios for a target}
}
\value{
The external error
}
\description{
Calculate external (deviation of runs) error for a target
}
