% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{normRunIntErr}
\alias{normRunIntErr}
\title{Calculate internal (counting) error for a target.}
\usage{
normRunIntErr(counts)
}
\arguments{
\item{counts}{A vector of counts for each run of a target.}
}
\value{
The internal error
}
\description{
Calculate internal (counting) error for a target.
}
