% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{numRun}
\alias{numRun}
\title{Count runs}
\usage{
numRun(from, to = "present", sys = "both")
}
\arguments{
\item{from}{A date in character form}

\item{to}{A date in character. Defaults to present.}

\item{sys}{System- "USAMS", "CFAMS", defaults to both.}
}
\value{
A list: number of runs, number of wheels
}
\description{
Count runs
}
