% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{rcage}
\alias{rcage}
\title{Calculate radiocarbon years}
\usage{
rcage(fm)
}
\arguments{
\item{fm}{Fraction modern.}
}
\value{
radiocarbon age in years.
}
\description{
Calculate radiocarbon age from Fm
}
