% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_fudger}
\alias{read_fudger}
\title{Read Fudger or SNICSer Output}
\usage{
read_fudger(resfile)
}
\arguments{
\item{resfile}{character. A SNICSer format file with path.}
}
\value{
A list of data tables for each chunk in file.
}
\description{
Reads normalized and blank corrected data in Fudger format
from Fudger or SNICSer blank correction
"print" output. Files should be in tsv format with standard SNICS
headers.
}
