% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{readResfile}
\alias{readResfile}
\title{Read raw results data from file}
\usage{
readResfile(file)
}
\arguments{
\item{file}{A file path to a SNICS format results file.}
}
\value{
A data frame of the wheel results.
}
\description{
Read raw results data from file
}
