% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{readSnicsfile}
\alias{readSnicsfile}
\title{Read SNICSer Output}
\usage{
readSnicsfile(wheel)
}
\arguments{
\item{file}{character. A SNICSer format file with path.}
}
\value{
A list of data tables for each chunk in file.
}
\description{
Reads normalized and blank corrected data from SNICSer blank correction
"print" output. Files should be in tsv format with standard SNICS
headers.
}
