% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{readWheelfile}
\alias{readWheelfile}
\title{Read a SNICS wheelfile}
\usage{
readWheelfile(wheel)
}
\arguments{
\item{wheel}{character. A USAMS format wheelfile/runlist.}
}
\value{
A data frame containing the wheelfile
}
\description{
Won't work for CFAMS wheels
}
