% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{recReport}
\alias{recReport}
\title{Rec_num report}
\usage{
recReport(rec)
}
\arguments{
\item{rec}{Receipt number to query.}
}
\value{
A html report in the current directory.
}
\description{
Produces a report of all runs of a rec_num.
}
