% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove outliers by interquartile range}
\usage{
removeOutliers(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
The vector with outliers replaced by `NA`
}
\description{
This will return a vector without points outside of 3 * IQR
}
