% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{rsd}
\alias{rsd}
\title{Relative SD}
\usage{
rsd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The relative SD of x
}
\description{
Provides the standard deviation of a vector normalized to the mean of the vector.
}
