% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{totErr}
\alias{totErr}
\title{Total error.}
\usage{
totErr(targErr, intrErr)
}
\arguments{
\item{targErr}{Measurement error.}

\item{intrErr}{Intrinsic error.}
}
\description{
Calculate the predicted variability in a population given a measurement
error and an intrisic error.
}

