% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_snics_results}
\alias{write_snics_results}
\title{Write out tsv file in SNICS results format}
\usage{
write_snics_results(snics_df, filename)
}
\arguments{
\item{snics_df}{A dataframe in SNICS_results format}

\item{filename}{Filename to write to.}
}
\value{

}
\description{
Write out tsv file in SNICS results format
}
