% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{conMICADAS}
\alias{conMICADAS}
\title{Open MICADAS DB connection}
\usage{
conMICADAS(database = "db_ac14")
}
\arguments{
\item{database}{Database to connect to. Defaults to "db_ac14".}
}
\value{
A MariaDB connection object
}
\description{
Uses credentials for database from system key store.
}
\seealso{
\code{\link{store_credentials}} to store user credentials
in the system key store.
}
