% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{conNOSAMS}
\alias{conNOSAMS}
\title{Open NOSAMS DB connection}
\usage{
conNOSAMS(database = "nosams-prod")
}
\arguments{
\item{database}{Database to connect to. Defaults to "nosams-prod".}
}
\value{
An odbc db connection object
}
\description{
Uses credentials for database from system key store if found,
takes a connection string from the CONSTRING environment variable
if not found in key store.
}
\seealso{
\code{\link{store_credentials}} to store user credentials
in the system key store.
}
