% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{doMBC}
\alias{doMBC}
\title{Apply mass balance blank correction}
\usage{
doMBC(fmmeas, fmblank, massmeas, massblank)
}
\arguments{
\item{fmmeas}{Normalized Fm of the sample(s)}

\item{fmblank}{Normalized Fm of the blank}

\item{massmeas}{Mass of the sample(s)}

\item{massblank}{Mass of the blank}
}
\value{
A vector of mass balance corrected Fm.
}
\description{
Note, use total mass - blank mass for mass of the sample
if appropriate
}
