% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.R
\name{getQCData}
\alias{getQCData}
\title{Get and format standards data}
\usage{
getQCData(
  from,
  to,
  sys = "both",
  useQC = FALSE,
  intcal = FALSE,
  getcurrents = FALSE,
  ...
)
}
\arguments{
\item{from}{Character vector of date in form 'YYYY-MM-DD'.}

\item{to}{Character vector of date in form 'YYYY-MM-DD'.}

\item{sys}{Character vector of system name: 'cfams', 'usams', or 'both'.}

\item{useQC}{Use data from QC table if true.}

\item{intcal}{Use standards data from intcal table if true.}

\item{getcurrents}{get current and count data if true.}

\item{...}{Additional options passed to getStandards}
}
\value{
A data frame of munged standards data
}
\description{
Get and format standards data
}
