% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getStandards}
\alias{getStandards}
\title{Get Standards}
\usage{
getStandards(from, to = "present", sys = "both", getcurrents = TRUE,
  rec = NULL)
}
\arguments{
\item{from}{Character vector of date in form 'YYYY-MM-DD'.}

\item{to}{Character vector of date in form 'YYYY-MM-DD'.}

\item{sys}{Character vector of system name: 'cfams', 'usams', or 'both'.}

\item{getcurrents}{logical. Get current and count data?}

\item{rec}{numeric. If supplied, get all records matching this receipt number.}
}
\description{
Get standards from database using standards table.
}
