% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getStandards}
\alias{getStandards}
\title{Get Standards}
\usage{
getStandards(
  from,
  to,
  sys = "both",
  getcurrents = TRUE,
  rec = NULL,
  osg = NULL
)
}
\arguments{
\item{from}{Character vector of date in form 'YYYY-MM-DD'.}

\item{to}{Character vector of date in form 'YYYY-MM-DD'.}

\item{sys}{Character vector of system name: 'cfams', 'usams', or 'both'.}

\item{getcurrents}{logical. Get current and count data?}

\item{rec}{Numeric vector of NOSAMS reciept numbers. If supplied, get
all records matching these receipt numbers.}

\item{osg}{Numeric vector of NOSAMS osg numbers. Supersedes 'rec' parameter.}
}
\value{
A data frame of standards data
}
\description{
Get standards from database using standards table.
}
