% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{getWheel}
\alias{getWheel}
\title{Get results for a list of wheel names.}
\usage{
getWheel(wheel, test = FALSE)
}
\arguments{
\item{wheel}{A vector of wheel names in character format}

\item{test}{Get data from snics_results_test if true}
}
\value{
A data frame of analysed data
}
\description{
Get results for a list of wheel names.
}
