% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{intrErr}
\alias{intrErr}
\title{Intrinsic Error}
\usage{
intrErr(x, err, ...)
}
\arguments{
\item{x}{A vector of measured values.}

\item{err}{A vector of measurement errors.}

\item{...}{Additional parameters passed to sd.}
}
\value{
A vector of residual errors
}
\description{
Calculate the "extra" (intrinsic, residual) error.
}
