% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{numRun}
\alias{numRun}
\title{Count wheels and runs in a time period.}
\usage{
numRun(from, to, sys = "both")
}
\arguments{
\item{from}{A date object or date in character form (m-d-Y)}

\item{to}{A date object or date in character. Defaults to present.}

\item{sys}{System- "USAMS", "CFAMS", defaults to both.}
}
\value{
A list: number of runs, number of wheels
}
\description{
Uses target runtimes from snics_results for data source.
}
