% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_bats}
\alias{read_bats}
\title{Read BATS format MICADAS output from Excel files}
\usage{
read_bats(file)
}
\arguments{
\item{file}{Path to BATS file}
}
\value{
A parsed dataframe of MICADAS results
}
\description{
Read BATS format MICADAS output from Excel files
}
