% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove outliers by interquartile range}
\usage{
removeOutliers(x, multiplier = 1.5)
}
\arguments{
\item{x}{A vector of numeric values}

\item{multiplier}{A value to multiply by IQR}
}
\value{
The vector with outliers replaced by `NA`
}
\description{
Outliers are defined as points more than a multiplier * IQR above the
third quartile or multiplier * IQR below the first quartile. The
default multiplier is 1.5.
}
