% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{sigma}
\alias{sigma}
\title{Sigma}
\usage{
sigma(fmm, fmc, em, ec = 0)
}
\arguments{
\item{fmm}{Measured value.}

\item{fmc}{Expected value.}

\item{em}{Measurement error.}

\item{ec}{Error of expected value.}
}
\value{
A vector of sigma values
}
\description{
Also known as a z-score. Provides an estimate of the quality of
a measurement relative to it's expected value.
}
