% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{sumWheelStdsSR}
\alias{sumWheelStdsSR}
\title{Wheel standards summary}
\usage{
sumWheelStdsSR(wheel)
}
\arguments{
\item{wheel}{Wheel to summarize.}
}
\value{
A dataframe with summary information
}
\description{
Produces a report of all standards on a wheel
}
