% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{totErr}
\alias{totErr}
\title{Total error.}
\usage{
totErr(rep_err, res_err, fm)
}
\arguments{
\item{rep_err}{Measurement error.}

\item{res_err}{Intrinsic error.}

\item{fm}{Fraction modern (optional).}
}
\value{
A vector of total errors.
}
\description{
Calculate the predicted variability in a population given a measurement
error and an intrisic error.
}
